This is needed

+++ SQL files/US_Neighborhood_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE US_NEIGHBORHOOD_DATA (
     City VARCHAR(500) NOT NULL,
     Locality VARCHAR(500) NOT NULL,
-    CostOfLiving DECIMAL(5,2),
+    CostOfLiving INT,
     CrimeRate INT,
     PublicTransportation INT,
     DistanceToWork INT,
